fix: updates file paths after folder restructuring and creates modules
